Successfully evaluate Let statements
For let statements, we have to evaluate the value-producing expression
and also evaluate an identifier that's bound to a name. And also we
should get an error when we try to evaluate an unbound identifier. So we
make environment hash map to keep track of variables and their
associated values.